SimplifiedHash class and WordExplorer algo are implemented. All double changed to floats. Much faster hash generated and much smaller in the memory.
Former-commit-id: 46a60b0441e87e9cc749de39116e0236c4f14e54

+++ src/core/PProbas.java
@@ -16,26 +16,26 @@ public class PProbas implements Serializable {
     
     private static final long serialVersionUID = 3000L;
     
-    double[][][] pp=null; //[nodeIndex][site][state]
+    float[][][] pp=null; //[nodeIndex][site][state]
     //note that the index of a node is defined through its DFS pre_order
 
     public PProbas(int nodeCount, int siteCount, int stateCount) {
-        pp=new double[nodeCount][siteCount][stateCount];
+        pp=new float[nodeCount][siteCount][stateCount];
     }
     
-    public void setStates(int nodeId, int site, double[] states) {
+    public void setStates(int nodeId, int site, float[] states) {
         for (int i = 0; i < pp[0][0].length; i++) {
             pp[nodeId][site][i]=states[i];
         }
     }
-    public void setState(int nodeId, int site, int stateIndex, double ppValue ) {
+    public void setState(int nodeId, int site, int stateIndex, float ppValue ) {
         pp[nodeId][site][stateIndex]=ppValue;
     }    
     public double getPP(int nodeId, int site, int state) {
         return pp[nodeId][site][state];
     }
     
-    public double[][] getPPSet(int nodeId, int siteStart, int siteEnd) {
+    public float[][] getPPSet(int nodeId, int siteStart, int siteEnd) {
         if (siteStart>siteEnd) {return null;}
         return Arrays.copyOfRange(pp[nodeId], siteStart, siteEnd+1);
     }
